+++ src/Services/LoggerService/LoggerService.h
@@ -44,7 +44,7 @@ namespace Mengine
         bool validMessage( ELoggerLevel _level, uint32_t _flag ) const override;
 
     public:
-        void logMessage( ELoggerLevel _level, uint32_t _flag, const Char * _message, uint32_t _size ) override;
+        void logMessage( ELoggerLevel _level, uint32_t _flag, const Char * _message, size_t _size ) override;
         uint32_t getCountMessage( ELoggerLevel _level ) override;
 
     public:
